Standard deviation.
The standard deviation of a random variable, sample, statistical population, data set, or probability distribution is the square root of its variance. It is algebraically simpler, though in practice, less robust than the average absolute deviation.[3][4] A useful property of the standard deviation is that unlike the variance, it is expressed in the same unit as the data.